add metrics.
