+++ pkg/sink/codec/common/compress.go
@@ -1,11 +1,12 @@
 package common
 
 import (
+	"github.com/pingcap/tiflow/cdc/model"
 	"github.com/pingcap/tiflow/pkg/compression"
 	"github.com/pingcap/tiflow/pkg/errors"
 )
 
-func Compress(cc string, data []byte) ([]byte, error) {
+func Compress(changefeedID model.ChangeFeedID, cc string, data []byte) ([]byte, error) {
 	oldSize := len(data)
 	compressed, err := compression.Encode(cc, data)
 	if err != nil {
@@ -13,7 +14,9 @@ func Compress(cc string, data []byte) ([]byte, error) {
 	}
 
 	newSize := len(compressed)
-	_ = float64(oldSize) / float64(newSize) * 100
+	ratio := float64(oldSize) / float64(newSize) * 100
+
+	compressionRatio.WithLabelValues(changefeedID.Namespace, changefeedID.ID).Observe(ratio)
 
 	return compressed, nil
 }
